Split out time panel module
